Align error title property to "errorTitle"
In the cq:template file and the share.html, a reference is made to the "errorTitle" property, while the value in the dialog is "./errorText".
All values are prefilled in the dialog, except for the error title. Making adjustments to this field in the dialog will not impact the value in the jcr since it is storing its value in "./errorText" instead of "./errorTitle".

Given the standard naming convention (modalTitle, successTitle,...) I aligned the property name to errorTitle. No change to the cq:template or html is required since they already use errorTitle.